Create OddsAndEvens
For this project you are going to program a game called "Odds and Evens". The game is similar to rock paper scissors. It is played between two players, in your version it will be you versus the computer. Each player will choose either "odds" or "evens", since you’re playing the computer you will get first pick. Once you have chosen your side, you each choose a number of fingers to play- 0 to 5. The winner is determined by whether the sum of your fingers is odd or even (depending on what you chose). Here's a clip of the game being played: https://youtu.be/4ZOLs03vILs?t=1m

For example, let’s say you choose "evens". That means the computer gets "odds". You choose to play 2 fingers and the computer plays 3. 2 + 3 = 5, which is odd so the computer wins.